ingestion/requirements.txt
openai>=1.12.0
numpy>=1.26.0
faiss-cpu>=1.7.4
tenacity>=8.2.0
